Update counting code

+++ events/guild/messageCreate.js
@@ -1,25 +1,13 @@
 const { pgClient } = require("../../main");
 const fs = require("fs");
 var Filter = require('bad-words');
+var botMsg = false;
 
 module.exports = async (Discord, client, message) => {
-    if (message.author.bot) {
+    if (message.author.bot && message.channel.id!="974471842704277524") {
         return;
     }   
 
-    if (message.content.toLowerCase().startsWith("im ")) {
-        if(parseInt(Math.random()*10)==2){
-            await message.channel.send(`Hi ${message.content.substring(3)}, I'm dad!`);
-        }
-    } else if (message.content.toLowerCase().startsWith("i'm ")) {
-        if(parseInt(Math.random()*10)==2){
-            await message.channel.send(`Hi ${message.content.substring(4)}, I'm dad!`);
-        }
-    } else if ( message.content.toLowerCase().startsWith("i am ")){
-        if(parseInt(Math.random()*10)==2){
-            await message.channel.send(`Hi ${message.content.substring(5)}, I'm dad!`);
-        }
-    }
     let filter = new Filter();
     let rawData = fs.readFileSync('./config.json');
     let config = JSON.parse(rawData);
@@ -33,7 +21,7 @@ module.exports = async (Discord, client, message) => {
     }
     filter.addWords(...censor);
     filter.removeWords(...uncensor);
-    if (filter.isProfane(message.content.toLowerCase())){
+    if (filter.isProfane(message.content.toLowerCase()) && message.channel.id!="925454253319921695"){
         let member = message.author.id.toString();
         await message.delete();
         if (message.content.toLowerCase().includes("owo") || message.content.toLowerCase().includes("owo")) {
@@ -46,18 +34,15 @@ module.exports = async (Discord, client, message) => {
     if(message.content.toLowerCase().includes('gp')){
         message.channel.send("Who are you, Tony?");
     }
-    if(parseInt(Math.random()*100)==2){
-        if(message.channel.id==853288359505821716 || message.channel.id==904848756090998834){
-            message.channel.send("no one asked?");
-        }
-    }
+
     if (message.channel.id == config.countingchannel[0].channel_id) { // this needs to be fetched from config.json later
         pgClient.query(`SELECT * FROM counting`, async (err, res) => {
-            if (message.content.startsWith(parseInt(res.rows[0].number)+1) && message.author.id != res.rows[0].user_id) { // might not work idk if the [0] should be there :clown:
-                pgClient.query(`UPDATE counting SET number = ('${parseInt(res.rows[0].number)+1}'), user_id = ('${message.author.id}')`);
-            } else {
-                await message.delete();
-            }
+                if (message.content.startsWith(parseInt(res.rows[0].number)+1) && message.author.id != res.rows[0].user_id) { // might not work idk if the [0] should be there :clown:
+                    pgClient.query(`UPDATE counting SET number = ('${parseInt(res.rows[0].number)+1}'), user_id = ('${message.author.id}')`);
+                } else {
+                    await message.delete();
+                }
+
         })
     }
     if (!message.content.startsWith(prefix) || message.author.bot) return;
